requirements.txt
asyncio
websockets
nest_asyncio
Pillow
opencv-python-headless